refactor: split the code into modules
