+++ service/index.ts
@@ -1,56 +1,11 @@
-import { $ } from 'bun'
-import { arch, platform } from 'os'
-import { Elysia } from 'elysia'
-import { execFile } from 'child_process'
-import { staticPlugin } from '@elysiajs/static'
 import { cors } from '@elysiajs/cors'
 import { cron, Patterns } from '@elysiajs/cron'
-import { drizzle } from 'drizzle-orm/mysql2'
-import mysql from 'mysql2/promise'
-import * as schema from './schema'
-import { count, eq, sql, sum } from 'drizzle-orm'
-
-const connection = await mysql.createConnection(
-  Bun.env.DATABASE_URL || 'mysql://root:password@localhost:3306/database',
-)
-
-const db = drizzle(connection, { schema, mode: 'default' })
-
-export type RunVlmcsType = {
-  host: string
-  port?: number
-  app?: number
-  protocol?: number
-}
-
-const runVlmcs = ({
-  host,
-  port = 1688,
-  app = 26,
-  protocol = 6,
-}: RunVlmcsType) => {
-  return new Promise<{
-    content: string
-    delay: number
-    host: string
-    status: boolean
-  }>(resolve => {
-    const before = Date.now()
-    execFile(
-      `./service/binaries/vlmcs-${platform()}-${arch()}`,
-      [`${host}:${port}`, `-${protocol}`, `-l ${app}`],
-      { timeout: 10 * 1000 },
-      (err, stdout) => {
-        resolve({
-          host,
-          delay: Date.now() - before,
-          content: stdout.trim(),
-          status: err ? false : true,
-        })
-      },
-    )
-  })
-}
+import { staticPlugin } from '@elysiajs/static'
+import { Elysia } from 'elysia'
+import { db, runCheck } from './db'
+import type { RunVlmcsParams } from './types'
+import { runVlmcs } from './utils'
+import './vlmcsd'
 
 const app = new Elysia()
 
@@ -67,55 +22,14 @@ app.use(
   cron({
     name: 'heartbeat',
     pattern: Patterns.everyMinutes(10),
-    async run() {
-      const servers = await db.query.server.findMany()
-      if (Array.isArray(servers) && servers?.length > 0) {
-        for await (const item of servers) {
-          const result = await runVlmcs({
-            host: item.host,
-            port: item.port,
-          })
-          await db.insert(schema.logs).values({
-            ...result,
-            createdAt: new Date(),
-          })
-        }
-
-        const result = await db
-          .select({
-            host: schema.logs.host,
-            total: count(),
-            success: sum(eq(schema.logs.status, true)).mapWith(Number),
-            fail: sum(eq(schema.logs.status, false)).mapWith(Number),
-            delay: sql<number>`avg(case when ${schema.logs.status} = true then ${schema.logs.delay} else null end)`,
-          })
-          .from(schema.logs)
-          .groupBy(schema.logs.host)
-
-        for await (const item of result) {
-          const { host, total, success, fail, delay } = item
-
-          await db
-            .update(schema.server)
-            .set({
-              total,
-              success,
-              fail,
-              delay: Number(delay),
-              rate: Number((success / total).toFixed(2)),
-              updatedAt: new Date(),
-            })
-            .where(eq(schema.server.host, host))
-        }
-      }
-    },
+    run: runCheck,
   }),
 )
 
 app.get('/*', () => Bun.file('dist/index.html'))
 
 app.post('/api/check', async request => {
-  const body = request.body as RunVlmcsType
+  const body = request.body as RunVlmcsParams
   return await runVlmcs(body)
 })
 
@@ -126,19 +40,3 @@ app.get('/api/server', async () => {
 app.listen(Bun.env.PORT || 3000)
 
 console.log(`Elysia is running at on port ${app.server?.url} ...`)
-
-try {
-  platform() === 'win32'
-    ? await $`taskkill /IM vlmcsd* /F`.nothrow()
-    : await $`pkill -f vlmcsd`.nothrow()
-} catch (err) {
-  console.error(err)
-}
-
-const vlmcsd = Bun.spawnSync([
-  `./service/binaries/vlmcsd-${platform()}-${arch()}`,
-])
-
-if (vlmcsd.success) {
-  console.log('Vlmcsd has started')
-}
